+++ sdgym/__init__.py
@@ -8,16 +8,19 @@ __author__ = 'MIT Data To AI Lab'
 __copyright__ = 'Copyright (c) 2018, MIT Data To AI Lab'
 __email__ = 'dailabmit@gmail.com'
 __license__ = 'MIT'
-__version__ = '0.3.0'
+__version__ = '0.4.0.dev2'
 
-from sdgym import benchmark, results, synthesizers
+from sdgym import benchmark, synthesizers
 from sdgym.benchmark import run
+from sdgym.collect import collect_results
 from sdgym.datasets import load_dataset
+from sdgym.summary import make_summary_spreadsheet
 
 __all__ = [
     'benchmark',
     'synthesizers',
-    'results',
     'run',
     'load_dataset',
+    'collect_results',
+    'make_summary_spreadsheet'
 ]