+++ sdgym/benchmark.py
@@ -11,15 +11,17 @@ from pathlib import Path
 import compress_pickle
 import numpy as np
 import pandas as pd
+import torch
 import tqdm
 
 from sdgym.datasets import get_dataset_paths, load_dataset, load_tables
 from sdgym.errors import SDGymError
 from sdgym.metrics import get_metrics
 from sdgym.progress import TqdmLogger, progress
+from sdgym.s3 import is_s3_path, write_csv, write_file
 from sdgym.synthesizers.base import Baseline
 from sdgym.utils import (
-    build_synthesizer, format_exception, get_synthesizers_dict, import_object, used_memory)
+    build_synthesizer, format_exception, get_synthesizers, import_object, used_memory)
 
 LOGGER = logging.getLogger(__name__)
 
@@ -74,6 +76,7 @@ def _compute_scores(metrics, real_data, synthetic_data, metadata, output):
         try:
             LOGGER.info('Computing %s on dataset %s', metric_name, metadata._metadata['name'])
             score = metric.compute(*metric_args)
+            normalized_score = metric.normalize(score)
         except Exception:
             LOGGER.exception('Metric %s failed on dataset %s. Skipping.',
                              metric_name, metadata._metadata['name'])
@@ -81,13 +84,14 @@ def _compute_scores(metrics, real_data, synthetic_data, metadata, output):
 
         scores[-1].update({
             'score': score,
+            'normalized_score': normalized_score,
             'error': error,
             'metric_time': (datetime.utcnow() - start).total_seconds()
         })
         output['scores'] = scores  # re-inject list to multiprocessing output
 
 
-def _score(synthesizer, metadata, metrics, iteration, output=None):
+def _score(synthesizer, metadata, metrics, iteration, output=None, max_rows=None):
     if output is None:
         output = {}
 
@@ -96,7 +100,7 @@ def _score(synthesizer, metadata, metrics, iteration, output=None):
     output['timeout'] = True  # To be deleted if there is no error
     output['error'] = 'Load Timeout'  # To be deleted if there is no error
     try:
-        real_data = load_tables(metadata)
+        real_data = load_tables(metadata, max_rows)
 
         LOGGER.info('Running %s on %s dataset %s; iteration %s; %s',
                     name, metadata.modality, metadata._metadata['name'], iteration, used_memory())
@@ -153,7 +157,8 @@ def _run_job(args):
     # Reset random seed
     np.random.seed()
 
-    synthesizer, metadata, metrics, iteration, cache_dir, timeout, run_id = args
+    synthesizer, metadata, metrics, iteration, cache_dir, \
+        timeout, run_id, aws_key, aws_secret, max_rows = args
 
     name = synthesizer['name']
     dataset_name = metadata._metadata['name']
@@ -164,7 +169,7 @@ def _run_job(args):
     if timeout:
         output = _score_with_timeout(timeout, synthesizer, metadata, metrics, iteration)
     else:
-        output = _score(synthesizer, metadata, metrics, iteration)
+        output = _score(synthesizer, metadata, metrics, iteration, max_rows=max_rows)
 
     scores = output.get('scores')
     if not scores:
@@ -183,14 +188,16 @@ def _run_job(args):
         scores['error'] = output['error']
 
     if cache_dir:
-        base_path = str(cache_dir / f'{name}_{dataset_name}_{iteration}_{run_id}')
+        cache_dir_name = str(cache_dir)
+        base_path = f'{cache_dir_name}/{name}_{dataset_name}_{iteration}_{run_id}'
         if scores is not None:
-            scores.to_csv(base_path + '_scores.csv', index=False)
+            write_csv(scores, f'{base_path}_scores.csv', aws_key, aws_secret)
         if 'synthetic_data' in output:
-            compress_pickle.dump(output['synthetic_data'], base_path + '.data.gz')
+            synthetic_data = compress_pickle.dumps(output['synthetic_data'], compression='gzip')
+            write_file(synthetic_data, f'{base_path}.data.gz', aws_key, aws_secret)
         if 'exception' in output:
-            with open(base_path + '_error.txt', 'w') as error_file:
-                error_file.write(output['exception'])
+            exception = output['exception'].encode('utf-8')
+            write_file(exception, f'{base_path}_error.txt', aws_key, aws_secret)
 
     return scores
 
@@ -218,9 +225,10 @@ def _run_on_dask(jobs, verbose):
     return dask.compute(*persisted)
 
 
-def run(synthesizers, datasets=None, datasets_path=None, modalities=None, bucket=None,
+def run(synthesizers=None, datasets=None, datasets_path=None, modalities=None, bucket=None,
         metrics=None, iterations=1, workers=1, cache_dir=None, show_progress=False,
-        timeout=None, output_path=None):
+        timeout=None, output_path=None, aws_key=None, aws_secret=None, jobs=None,
+        max_rows=None, max_columns=None):
     """Run the SDGym benchmark and return a leaderboard.
 
     The ``synthesizers`` object can either be a single synthesizer or, an iterable of
@@ -251,20 +259,21 @@ def run(synthesizers, datasets=None, datasets_path=None, modalities=None, bucket
         bucket (str):
             Name of the bucket from which the datasets must be downloaded if not found locally.
         iterations (int):
-            Number of iterations to perform over each dataset and synthesizer. Defaults to 3.
+            Number of iterations to perform over each dataset and synthesizer. Defaults to 1.
         workers (int or str):
             If ``workers`` is given as an integer value other than 0 or 1, a multiprocessing
             Pool is used to distribute the computation across the indicated number of workers.
-            If the string ``dask`` is given, the computation is distributed using ``dask``.
-            In this case, setting up the ``dask`` cluster and client is expected to be handled
-            outside of this function.
+            If ``workers`` is -1, the number of workers will be automatically determined by
+            the number of GPUs (if available) or the number of CPU cores. If the string ``dask``
+            is given, the computation is distributed using ``dask``. In this case, setting up the
+            ``dask`` cluster and client is expected to be handled outside of this function.
         cache_dir (str):
             If a ``cache_dir`` is given, intermediate results are stored in the indicated directory
             as CSV files as they get computted. This allows inspecting results while the benchmark
             is still running and also recovering results in case the process does not finish
             properly. Defaults to ``None``.
         show_progress (bool):
-            Whether to use tqdm to keep track of the progress. Defaults to ``True``.
+            Whether to use tqdm to keep track of the progress. Defaults to ``False``.
         timeout (int):
             Maximum number of seconds to wait for each dataset to
             finish the evaluation process. If not passed, wait until
@@ -273,55 +282,95 @@ def run(synthesizers, datasets=None, datasets_path=None, modalities=None, bucket
             If an ``output_path`` is given, the generated leaderboard will be stored in the
             indicated path as a CSV file. The given path must be a complete path including
             the ``.csv`` filename.
+        aws_key (str):
+            If an ``aws_key`` is provided, the given access key id will be used to read
+            from the specified bucket.
+        aws_secret (str):
+            If an ``aws_secret`` is provided, the given secret access key will be used to read
+            from the specified bucket.
+        jobs (list[tuple]):
+            List of jobs to execute, as a sequence of tuple-like objects containing synthesizer,
+            dataset and iteration-id specifications. If not passed, the jobs list is build by
+            combining the synthesizers and datasets given instead.
+        max_rows (int):
+            Cap the number of rows to model from each dataset. Rows will be selected in order.
+        max_columns (int):
+            Cap the number of columns to model from each dataset.
+            Columns will be selected in order.
 
     Returns:
         pandas.DataFrame:
             A table containing one row per synthesizer + dataset + metric + iteration.
     """
-    synthesizers = get_synthesizers_dict(synthesizers)
-    datasets = get_dataset_paths(datasets, datasets_path, bucket)
-    run_id = os.getenv('RUN_ID') or str(uuid.uuid4())[:10]
-
-    if cache_dir:
+    if cache_dir and not is_s3_path(cache_dir):
         cache_dir = Path(cache_dir)
         os.makedirs(cache_dir, exist_ok=True)
 
-    jobs = list()
-    for dataset in datasets:
-        metadata = load_dataset(dataset)
-        for synthesizer in synthesizers:
-            modalities_ = modalities or synthesizer.get('modalities')
-            if not modalities_ or metadata.modality in modalities_:
+    run_id = os.getenv('RUN_ID') or str(uuid.uuid4())[:10]
+
+    if workers == -1:
+        if torch.cuda.is_available():
+            workers = torch.cuda.device_count()
+        else:
+            workers = multiprocessing.cpu_count()
+
+    if jobs is None:
+        synthesizers = get_synthesizers(synthesizers)
+        datasets = get_dataset_paths(datasets, datasets_path, bucket, aws_key, aws_secret)
+
+        job_tuples = list()
+        for dataset in datasets:
+            for synthesizer in synthesizers:
                 for iteration in range(iterations):
-                    args = (
-                        synthesizer,
-                        metadata,
-                        metrics,
-                        iteration,
-                        cache_dir,
-                        timeout,
-                        run_id,
-                    )
-                    jobs.append(args)
+                    job_tuples.append((synthesizer, dataset, iteration))
+
+    else:
+        job_tuples = list()
+        for synthesizer, dataset, iteration in jobs:
+            job_tuples.append((
+                get_synthesizers([synthesizer])[0],
+                get_dataset_paths([dataset], datasets_path, bucket, aws_key, aws_secret)[0],
+                iteration
+            ))
+
+    job_args = list()
+    for synthesizer, dataset, iteration in job_tuples:
+        metadata = load_dataset(dataset, max_columns=max_columns)
+        modalities_ = modalities or synthesizer.get('modalities')
+        if not modalities_ or metadata.modality in modalities_:
+            args = (
+                synthesizer,
+                metadata,
+                metrics,
+                iteration,
+                cache_dir,
+                timeout,
+                run_id,
+                aws_key,
+                aws_secret,
+                max_rows,
+            )
+            job_args.append(args)
 
     if workers == 'dask':
-        scores = _run_on_dask(jobs, show_progress)
+        scores = _run_on_dask(job_args, show_progress)
     else:
         if workers in (0, 1):
-            scores = map(_run_job, jobs)
+            scores = map(_run_job, job_args)
         else:
             pool = concurrent.futures.ProcessPoolExecutor(workers)
-            scores = pool.map(_run_job, jobs)
+            scores = pool.map(_run_job, job_args)
 
-        scores = tqdm.tqdm(scores, total=len(jobs), file=TqdmLogger())
+        scores = tqdm.tqdm(scores, total=len(job_args), file=TqdmLogger())
         if show_progress:
-            scores = tqdm.tqdm(scores, total=len(jobs))
+            scores = tqdm.tqdm(scores, total=len(job_args))
 
     if not scores:
         raise SDGymError("No valid Dataset/Synthesizer combination given")
 
     scores = pd.concat(scores)
+
     if output_path:
-        scores.to_csv(output_path, index=False)
+        write_csv(scores, output_path, aws_key, aws_secret)
 
     return scores