+++ sdgym/datasets.py
@@ -1,13 +1,15 @@
 import io
+import itertools
+import json
 import logging
-import urllib.request
 from pathlib import Path
-from xml.etree import ElementTree
 from zipfile import ZipFile
 
 import appdirs
 import pandas as pd
 from sdv import Metadata
+
+from sdgym.s3 import get_s3_client
 
 LOGGER = logging.getLogger(__name__)
 
@@ -17,21 +19,21 @@ BUCKET_URL = 'https://{}.s3.amazonaws.com/'
 TIMESERIES_FIELDS = ['sequence_index', 'entity_columns', 'context_columns', 'deepecho_version']
 
 
-def download_dataset(dataset_name, datasets_path=None, bucket=None):
+def download_dataset(dataset_name, datasets_path=None, bucket=None, aws_key=None, aws_secret=None):
     datasets_path = datasets_path or DATASETS_PATH
     bucket = bucket or BUCKET
-    url = BUCKET_URL.format(bucket) + f'{dataset_name}.zip'
 
-    LOGGER.info('Downloading dataset %s from %s', dataset_name, url)
-    response = urllib.request.urlopen(url)
-    bytes_io = io.BytesIO(response.read())
+    LOGGER.info('Downloading dataset %s from %s', dataset_name, bucket)
+    s3 = get_s3_client(aws_key, aws_secret)
+    obj = s3.get_object(Bucket=bucket, Key=f'{dataset_name}.zip')
+    bytes_io = io.BytesIO(obj['Body'].read())
 
     LOGGER.info('Extracting dataset into %s', datasets_path)
     with ZipFile(bytes_io) as zf:
         zf.extractall(datasets_path)
 
 
-def _get_dataset_path(dataset, datasets_path, bucket=None):
+def _get_dataset_path(dataset, datasets_path, bucket=None, aws_key=None, aws_secret=None):
     dataset = Path(dataset)
     if dataset.exists():
         return dataset
@@ -41,13 +43,37 @@ def _get_dataset_path(dataset, datasets_path, bucket=None):
     if dataset_path.exists():
         return dataset_path
 
-    download_dataset(dataset, datasets_path, bucket=bucket)
+    download_dataset(dataset, datasets_path, bucket=bucket, aws_key=aws_key, aws_secret=aws_secret)
     return dataset_path
 
 
-def load_dataset(dataset, datasets_path=None, bucket=None):
-    dataset_path = _get_dataset_path(dataset, datasets_path, bucket)
-    metadata = Metadata(str(dataset_path / 'metadata.json'))
+def _apply_max_columns_to_metadata(metadata, max_columns):
+    tables = metadata['tables']
+    for table in tables.values():
+        fields = table['fields']
+        if len(fields) > max_columns:
+            fields = dict(itertools.islice(fields.items(), max_columns))
+            table['fields'] = fields
+
+        structure = table.get('structure')
+        if structure:
+            structure['structure'] = structure['structure'][:max_columns]
+            structure['states'] = structure['states'][:max_columns]
+
+
+def load_dataset(dataset, datasets_path=None, bucket=None, aws_key=None, aws_secret=None,
+                 max_columns=None):
+    dataset_path = _get_dataset_path(dataset, datasets_path, bucket, aws_key, aws_secret)
+    with open(dataset_path / 'metadata.json') as metadata_file:
+        metadata_content = json.load(metadata_file)
+
+    if max_columns:
+        if len(metadata_content['tables']) > 1:
+            raise ValueError('max_columns is not supported for multi-table datasets')
+
+        _apply_max_columns_to_metadata(metadata_content, max_columns)
+
+    metadata = Metadata(metadata_content, dataset_path)
     tables = metadata.get_tables()
     if not hasattr(metadata, 'modality'):
         if len(tables) > 1:
@@ -69,9 +95,13 @@ def load_dataset(dataset, datasets_path=None, bucket=None):
     return metadata
 
 
-def load_tables(metadata):
+def load_tables(metadata, max_rows=None):
+    if max_rows and len(metadata.get_tables()) > 1:
+        raise ValueError('max_rows is not supported for multi-table datasets')
+
     real_data = metadata.load_tables()
     for table_name, table in real_data.items():
+        table = table.head(max_rows)
         fields = metadata.get_fields(table_name)
         columns = [
             column
@@ -83,14 +113,13 @@ def load_tables(metadata):
     return real_data
 
 
-def get_available_datasets(bucket=None):
-    bucket_url = BUCKET_URL.format(bucket or BUCKET)
-    response = urllib.request.urlopen(bucket_url)
-    tree = ElementTree.fromstring(response.read())
+def get_available_datasets(bucket=None, aws_key=None, aws_secret=None):
+    s3 = get_s3_client(aws_key, aws_secret)
+    response = s3.list_objects(Bucket=bucket or BUCKET)
     datasets = []
-    for content in tree.findall('{*}Contents'):
-        key = content.find('{*}Key').text
-        size = int(content.find('{*}Size').text)
+    for content in response['Contents']:
+        key = content['Key']
+        size = int(content['Size'])
         if key.endswith('.zip'):
             datasets.append({
                 'name': key[:-len('.zip')],
@@ -118,7 +147,7 @@ def get_downloaded_datasets(datasets_path=None):
     return pd.DataFrame(datasets)
 
 
-def get_dataset_paths(datasets, datasets_path, bucket):
+def get_dataset_paths(datasets, datasets_path, bucket, aws_key, aws_secret):
     """Build the full path to datasets and ensure they exist."""
     if datasets_path is None:
         datasets_path = DATASETS_PATH
@@ -132,6 +161,6 @@ def get_dataset_paths(datasets, datasets_path, bucket):
             datasets = get_available_datasets()['name'].tolist()
 
     return [
-        _get_dataset_path(dataset, datasets_path, bucket)
+        _get_dataset_path(dataset, datasets_path, bucket, aws_key, aws_secret)
         for dataset in datasets
     ]