+++ sdgym/synthesizers/__init__.py
@@ -1,4 +1,5 @@
 from sdgym.synthesizers.clbn import CLBN
+from sdgym.synthesizers.gretel import Gretel, PreprocessedGretel
 from sdgym.synthesizers.identity import Identity
 from sdgym.synthesizers.independent import Independent
 from sdgym.synthesizers.medgan import MedGAN
@@ -9,6 +10,8 @@ from sdgym.synthesizers.sdv import (
 from sdgym.synthesizers.tablegan import TableGAN
 from sdgym.synthesizers.uniform import Uniform
 from sdgym.synthesizers.veegan import VEEGAN
+from sdgym.synthesizers.ydata import (
+    DRAGAN, WGAN_GP, PreprocessedDRAGAN, PreprocessedVanillaGAN, PreprocessedWGAN_GP, VanillaGAN)
 
 __all__ = (
     'CLBN',
@@ -25,4 +28,12 @@ __all__ = (
     'GaussianCopulaCategorical',
     'GaussianCopulaCategoricalFuzzy',
     'GaussianCopulaOneHot',
+    'Gretel',
+    'PreprocessedGretel',
+    'VanillaGAN',
+    'DRAGAN',
+    'WGAN_GP',
+    'PreprocessedDRAGAN',
+    'PreprocessedWGAN_GP',
+    'PreprocessedVanillaGAN',
 )