+++ sdgym/synthesizers/base.py
@@ -38,7 +38,7 @@ class Baseline:
 class SingleTableBaseline(Baseline):
     """Base class for all the SingleTable Baselines.
 
-    Sublcasses can choose to implement ``_fit_sample``, which will
+    Subclasses can choose to implement ``_fit_sample``, which will
     always be called with DataFrames and Table metadata dicts, or
     to overwrite the ``fit_sample`` method, which may be called with
     either DataFrames and Table dicts, or with dicts of tables and
@@ -46,15 +46,35 @@ class SingleTableBaseline(Baseline):
     """
 
     MODALITIES = ('single-table', )
+    CONVERT_TO_NUMERIC = False
+
+    def _transform_fit_sample(self, real_data, metadata):
+        ht = rdt.HyperTransformer()
+        columns_to_transform = list()
+        fields_metadata = metadata['fields']
+        id_fields = list()
+        for field in fields_metadata:
+            if fields_metadata.get(field).get('type') != 'id':
+                columns_to_transform.append(field)
+            else:
+                id_fields.append(field)
+
+        ht.fit(real_data[columns_to_transform])
+        transformed_data = ht.transform(real_data)
+        synthetic_data = self._fit_sample(transformed_data, metadata)
+        reverse_transformed_synthetic_data = ht.reverse_transform(synthetic_data)
+        reverse_transformed_synthetic_data[id_fields] = real_data[id_fields]
+        return reverse_transformed_synthetic_data
 
     def fit_sample(self, real_data, metadata):
+        _fit_sample = self._transform_fit_sample if self.CONVERT_TO_NUMERIC else self._fit_sample
         if isinstance(real_data, dict):
             return {
-                table_name: self._fit_sample(table, metadata.get_table_meta(table_name))
+                table_name: _fit_sample(table, metadata.get_table_meta(table_name))
                 for table_name, table in real_data.items()
             }
 
-        return self._fit_sample(real_data, metadata)
+        return _fit_sample(real_data, metadata)
 
 
 class MultiSingleTableBaseline(Baseline):